remote_graph_service/requirements.txt
fastapi[standard]>=0.115.11
faststream[rabbit]>=0.5.35
python-dotenv>=1.0.1
fastapi-utils>=0.8.0
typing_inspect>=0.9.0
scikit-learn>=1.6.1
umap-learn>=0.5.7
google-genai>=1.10.0